put snapshot test util into separate module
